core: refactoring bodies into separate files to add ellipsoid bodies
