+++ include/io_maps.hpp
@@ -1,7 +1,7 @@
 #ifndef IO_MAPS_HPP
 #define IO_MAPS_HPP
 
-#include <body.hpp>
+#include <body_container.hpp>
 #include <fiber_container_base.hpp>
 #include <fiber_container_finite_difference.hpp>
 #include <fiber_finite_difference.hpp>
